p4i integration Integrate 91139 from 3.1 to main: Nuke warning on windows.
[git-p4: depot-paths = "//depot/qt/main/": change = 91140]

+++ qmake/project.cpp
@@ -1008,9 +1008,9 @@ QMakeProject::doVariableReplace(QString &str, const QMap<QString, QStringList> &
 			end = arg_list[2].toInt();
 		    if(end < 0)
 			end += var.count() - 1;
-		    if(end < start)
+		    if(end < (int)start)
 			end = start;
-		    for(int i = start; i <= end && var.count() >= i; i++) {
+		    for(int i = start; i <= end && (int)var.count() >= i; i++) {
 			if(!replacement.isEmpty())
 			    replacement += " ";
 			replacement += var[i];
